requirements.txt
# Core
numpy>=1.25.0
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=2.0.1

# PDF / OCR
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.4.2
reportlab>=4.0.0
python-docx>=0.8.12
pdfplumber>=0.10.0

# Images
Pillow>=10.0.0
imageio>=2.30.1
opencv-python>=4.9.0

# Audio / Video
moviepy>=1.0.3
pydub>=0.25.1

# Text / NLP
nltk>=3.9.1
spacy>=3.6.1
gensim>=4.3.2

# Data Analysis / EDA
scipy>=1.12.0
scikit-learn>=1.3.2
statsmodels>=0.14.0

# JSON / YAML / Utilities
pyyaml>=7.0
tqdm>=4.66.0
requests>=2.31.0

# Data Generation / Synthetic
Faker>=23.3.0

# Optional / Extras
python-magic>=0.4.27
